Move configs.
